GitBase: Extract inner class 'GitWorkingTree'
Doing so for the same reason as for extracting
the 'WorkingTree' in a preceeding change.
