+++ downloader/src/main/kotlin/vcs/GitBase.kt
@@ -22,7 +22,6 @@ package com.here.ort.downloader.vcs
 import com.here.ort.downloader.VersionControlSystem
 import com.here.ort.downloader.WorkingTree
 import com.here.ort.utils.CommandLineTool
-import com.here.ort.utils.ProcessCapture
 
 import java.io.File
 import java.util.regex.Pattern
@@ -45,42 +44,5 @@ abstract class GitBase : VersionControlSystem(), CommandLineTool {
                 }
             }
 
-    open inner class GitWorkingTree(workingDir: File) : WorkingTree(workingDir, type) {
-        override fun isValid(): Boolean {
-            if (!workingDir.isDirectory) {
-                return false
-            }
-
-            // Do not use runGitCommand() here as we do not require the command to succeed.
-            val isInsideWorkTree = ProcessCapture(workingDir, "git", "rev-parse", "--is-inside-work-tree")
-            return isInsideWorkTree.isSuccess && isInsideWorkTree.stdout.trimEnd().toBoolean()
-        }
-
-        override fun isShallow(): Boolean {
-            val dotGitDir = run(workingDir, "rev-parse", "--absolute-git-dir").stdout.trimEnd()
-            return File(dotGitDir, "shallow").isFile
-        }
-
-        override fun getRemoteUrl() = run(workingDir, "remote", "get-url", getFirstRemote()).stdout.trimEnd()
-
-        override fun getRevision() = run(workingDir, "rev-parse", "HEAD").stdout.trimEnd()
-
-        override fun getRootPath() =
-                File(run(workingDir, "rev-parse", "--show-toplevel").stdout.trimEnd('\n', '/'))
-
-        private fun listRemoteRefs(namespace: String): List<String> {
-            val tags = run(workingDir, "ls-remote", "--refs", getFirstRemote(), "refs/$namespace/*").stdout.trimEnd()
-            return tags.lines().map {
-                it.split('\t').last().removePrefix("refs/$namespace/")
-            }
-        }
-
-        override fun listRemoteBranches() = listRemoteRefs("heads")
-
-        override fun listRemoteTags() = listRemoteRefs("tags")
-
-        private fun getFirstRemote() = run(workingDir, "remote", "show", "-n").stdout.lineSequence().first()
-    }
-
-    override fun getWorkingTree(vcsDirectory: File): WorkingTree = GitWorkingTree(vcsDirectory)
+    override fun getWorkingTree(vcsDirectory: File): WorkingTree = GitWorkingTree(vcsDirectory, this)
 }
